requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.7
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django_extensions==3.1.3
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.11.2
drf-jwt==1.19.1
factory-boy==3.2.0
Faker==8.12.1
gevent==21.8.0
greenlet==1.1.1
kombu==5.1.0
pika==1.2.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-logging-rabbitmq==2.0.0
pytz==2021.1
rich==10.7.0
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
vine==5.0.0
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
